Move helper function and remove debug output
